Multi cluster test support

+++ Vagrantfile.local.rb
@@ -13,39 +13,56 @@
 # file's provisioners will copy those settings into /etc/environment.
 #
 Vagrant.configure('2') do |config|
-  config.vm.synced_folder '/usr/local/share/ca-certificates',
-    '/usr/local/share/ca-certificates',
-    mount_options: ['ro']
 
-  debian_bundle_path = '/etc/ssl/certs/ca-certificates.crt'
-  redhat_bundle_path = '/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem'
+  # load any local SSL certificates if present
+  if Dir.exist?('/usr/local/share/ca-certificates')
+    debian_bundle_path = '/etc/ssl/certs/ca-certificates.crt'
+    redhat_bundle_path = '/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem'
 
-  config.vm.provision 'shell',  inline: <<-EOM.gsub(/^ {4}/,'')
-    DEBIAN_TOOL_PATH=/usr/sbin/update-ca-certificates
-    REDHAT_TOOL_PATH=/usr/bin/update-ca-trust
+    config.vm.synced_folder '/usr/local/share/ca-certificates',
+      '/usr/local/share/ca-certificates',
+      mount_options: ['ro']
 
-    # Debian/Ubuntu
-    if [[ -x $DEBIAN_TOOL_PATH ]]; then
-      SSL_PATH=#{debian_bundle_path}
-      $DEBIAN_TOOL_PATH
-    # RHEL/CentOS
-    elif [[ -x $REDHAT_TOOL_PATH ]]; then
-      for d in /usr/local/share/ca-certificates/*; do
-        for c in $d/*; do
-          ln -s "$c" /etc/pki/ca-trust/source/anchors
+    config.vm.provision 'shell',  inline: <<-EOM.gsub(/^ {4}/,'')
+      DEBIAN_TOOL_PATH=/usr/sbin/update-ca-certificates
+      REDHAT_TOOL_PATH=/usr/bin/update-ca-trust
+
+      # Debian/Ubuntu
+      if [[ -x $DEBIAN_TOOL_PATH ]]; then
+        SSL_PATH=#{debian_bundle_path}
+        $DEBIAN_TOOL_PATH
+      # RHEL/CentOS
+      elif [[ -x $REDHAT_TOOL_PATH ]]; then
+        for d in /usr/local/share/ca-certificates/*; do
+          for c in $d/*; do
+            ln -s "$c" /etc/pki/ca-trust/source/anchors
+          done
         done
-      done
 
-      SSL_PATH=#{redhat_bundle_path}
-      $REDHAT_TOOL_PATH enable
-      $REDHAT_TOOL_PATH extract
-    else
-      exit -1
-    fi
+        SSL_PATH=#{redhat_bundle_path}
+        $REDHAT_TOOL_PATH enable
+        $REDHAT_TOOL_PATH extract
+      else
+        exit -1
+      fi
 
-    echo "SSL_CERT_FILE=$SSL_PATH" >> \
-      /etc/environment
+      echo "SSL_CERT_FILE=$SSL_PATH" >> \
+        /etc/environment
 
+      umask 0277
+      echo 'Defaults env_keep += "SSL_CERT_FILE"' > \
+        /etc/sudoers.d/ssl
+    EOM
+
+    bundle_paths = [debian_bundle_path, redhat_bundle_path]
+
+    # Note: the box_download_ca_cert is used on the host, not the guest.
+    found_bps = bundle_paths.select { |bp| File.exist?(bp) }
+    config.vm.box_download_ca_cert = found_bps.first
+  end
+
+  # set proxies if present
+  config.vm.provision 'shell',  inline: <<-EOM.gsub(/^ {4}/,'')
     if [ -n "#{ENV['https_proxy']}" ]; then
       echo 'https_proxy=#{ENV['https_proxy']}' >> \
         /etc/environment
@@ -64,14 +81,5 @@ Vagrant.configure('2') do |config|
     umask 0277
     echo 'Defaults env_keep += "http_proxy https_proxy no_proxy"' > \
       /etc/sudoers.d/proxy
-    echo 'Defaults env_keep += "SSL_CERT_FILE"' > \
-      /etc/sudoers.d/ssl
   EOM
-
-  # Note: the box_download_ca_cert is used on the host, not the guest.
-  if File.exist?(debian_bundle_path)
-    config.vm.box_download_ca_cert = debian_bundle_path
-  elsif File.exist?(redhat_bundle_path)
-    config.vm.box_download_ca_cert = redhat_bundle_path
-  end
 end
